feat: enforce strict code quality with expanded linting and type checking (v0.14.0)
## Summary

- Expand Ruff linting rules (B, UP, SIM, RUF, C4, PIE, PT, ASYNC, S,
T20) and add mypy type checking with Pydantic plugin across all 134
source files
- Fix all 72 ruff lint violations across 36 files and all 267 mypy type
errors across 39 files
- Add mypy pre-commit hook (local, using `uv run mypy`) and CI
type-check job

## Test plan

- [x] `uvx ruff check src/ tests/` — All checks passed
- [x] `uvx ruff format --check src/ tests/` — 266 files already
formatted
- [x] `uv run mypy src/mcpbr/` — Success: no issues found in 134 source
files
- [x] `uv run pytest -m "not integration"` — 4293 passed, 0 failures
- [x] All pre-commit hooks pass (sync-version, ruff, ruff-format, mypy,
trailing-whitespace, end-of-file-fixer, check-yaml,
check-added-large-files, check-merge-conflict)

+++ src/mcpbr/analytics/statistical.py
@@ -279,15 +279,15 @@ def bootstrap_confidence_interval(
     # Generate bootstrap resamples
     bootstrap_means = []
     for _ in range(n_bootstrap):
-        resample = random.choices(values, k=n)
+        resample = random.choices(values, k=n)  # noqa: S311 -- not used for cryptographic purposes; statistical bootstrapping
         bootstrap_means.append(statistics.mean(resample))
 
     bootstrap_means.sort()
 
     # Percentile method
     alpha = 1.0 - confidence
-    lower_idx = int(math.floor((alpha / 2) * n_bootstrap))
-    upper_idx = int(math.floor((1.0 - alpha / 2) * n_bootstrap)) - 1
+    lower_idx = math.floor((alpha / 2) * n_bootstrap)
+    upper_idx = math.floor((1.0 - alpha / 2) * n_bootstrap) - 1
 
     # Clamp indices
     lower_idx = max(0, min(lower_idx, n_bootstrap - 1))
